requirements.txt
playwright
python-dotenv
rich
requests
pillow
